Hitrate 1, 0.881 för Bellffeature,OLD: hitrate 0.85, 0.66

+++ assignment_3/inl3_to_students/ocr_project/matlab/features2class.m
@@ -1,7 +1,7 @@
 function y = features2class(x,data)
 classVector = zeros(26,1);
     for(i = 1:26)
-        classVector(i) = predict(data{i},x');
+        classVector(i) = predict(data{i},double(x)');
     end
     y = find(classVector==1,1);
     if (isempty(y))
